refactor: extract certificate installation to separate plugin
