+++ lib/main.dart
@@ -1,7 +1,3 @@
-import 'dart:io';
-
-import 'package:flutter/foundation.dart';
-import 'package:flutter/services.dart';
 import 'package:go_transitions/go_transitions.dart';
 import 'package:hoyomi/apis/show_snack_bar.dart';
 import 'package:hoyomi/constraints/x_platform.dart';
@@ -9,6 +5,7 @@ import 'package:hoyomi/plugins/android_sdk_int.dart';
 import 'package:hoyomi/plugins/dot_env.dart';
 import 'package:hoyomi/plugins/firebase.dart';
 import 'package:hoyomi/plugins/init_services.dart';
+import 'package:hoyomi/plugins/install_cert.dart';
 import 'package:hoyomi/router/index.dart';
 
 import 'package:flutter/material.dart';
@@ -27,14 +24,11 @@ Future<void> main() async {
   await initializeDotEnv();
 
   await androidSdkIntInit();
-
-  if (!kIsWeb && androidSdkInt != null && androidSdkInt! < 25) {
-    await _installCert();
-  }
+  await initializeCert();
   await initializeFirebase();
   await initializeServices();
 
-  runApp(MainApp(androidSdkInt: androidSdkInt));
+  runApp(MainApp());
 
   // register background service
   // if (XPlatform.isAndroid || XPlatform.isIOS) {
@@ -43,30 +37,8 @@ Future<void> main() async {
   // }
 }
 
-class _MyHttpOverrides extends HttpOverrides {
-  @override
-  HttpClient createHttpClient(SecurityContext? context) {
-    return super.createHttpClient(context)
-      ..badCertificateCallback =
-          (X509Certificate cert, String host, int port) => true;
-  }
-}
-
-Future<void> _installCert() async {
-  ByteData data = await PlatformAssetBundle().load(
-    'assets/ca/lets-encrypt-r3.pem',
-  );
-  SecurityContext.defaultContext.setTrustedCertificatesBytes(
-    data.buffer.asUint8List(),
-  );
-
-  HttpOverrides.global = _MyHttpOverrides();
-}
-
 class MainApp extends StatelessWidget {
-  final int? androidSdkInt;
-
-  const MainApp({super.key, required this.androidSdkInt});
+  const MainApp({super.key});
 
   @override
   Widget build(BuildContext context) {
